requirements.txt
kraken
google-generativeai
lxml
reportlab
Pillow
PyYAML
python-dotenv
pypdf
